Added misc comments to classes and methods
Adding hacked together class to pull and process CSV for all user ratings. This StartupDebug class will be removed and it's behaviors split out into their appropriate classes.
Added user ratings to application data.

Changing the requestqueue to a concurrentqueue to handle cross thread calls. Was receiving an error that the queue was empty after checking to make sure it was not empty.

Added queue size to log window.

Tied a number in the main screen to show each loop of the call queue.